pyproject.toml
[tool.poetry]
name = "rag"
version = "0.1.0"
description = ""
authors = ["Tim Pap (@blackpr)"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
llama-index = "^0.9.30"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
